refactor: Players의 턴 반복 로직을 Board로 분리
